requirements.txt
discord.py==2.3.2
qrcode
Pillow
requests
google-generativeai
flask
